task3 update

+++ js/task-3.js
@@ -1,7 +1,8 @@
 function getElementWidth(content, padding, border) {
   const contentWidth = parseFloat(content);
-  const paddingWidth = parseFloat(padding);
-  const borderWidth = parseFloat(border);
+
+  const paddingWidth = parseInt(padding);
+  const borderWidth = parseInt(border);
   const totalWidth = contentWidth + 2 * (paddingWidth + borderWidth);
 
   return totalWidth;
